Raise the bank count to 32 and add bank navigation (firmware 0.11)
Banks go from 8 to 32. Navigation, since stepping through 32 banks one
at a time is not usable:
- A short press on Bank Up / Down steps one bank, a long press jumps
  Bank_Jump_Step banks (new global setting, default 8). Both wrap
  around at either end, which they did not before.
- A new Bank command type jumps to a given bank (GoTo) or moves
  relative to the current one (Up / Down), so one bank can act as a
  setlist index. The change is applied after the button's remaining
  commands have been sent, so the rest of the list still comes from the
  bank the button belongs to.

Firmware: the per-bank toggle, LED-toggle and long-press bitmasks widen
from 8 to 32 bits, the saved-state journal stores 32-bit masks (entry
grows to 66 bytes, new marker, so entries from older firmware are
ignored once), and the two bank switches get their own press timers.

Every configuration offset now derives from named sizes in
flash_midi_settings.h instead of a hardcoded 96 byte bank-string
region, with a _Static_assert that the total fits the reserved pages.
The hardcoded 96 was 8 banks x 12 bytes: with 32 banks it put the
commands, LED modes, labels and long press sets 288 bytes early, which
showed up on hardware as button labels and commands going missing while
bank names still read correctly.

Python/GUI: 32 banks everywhere, the Bank command type with an action
and target selector, Bank_Jump_Step in Global Settings, and padding so
a configuration that defines fewer banks (including one written for the
8 bank firmware) still flashes with the rest left empty. The
configurator always shows all 32 banks. The PC Bank Select field is
renamed BankSel in the editor to tell it apart from the Bank command.

Configuration grows to 22208 bytes over 11 flash pages and takes about
13 seconds to transfer. 18 tests, including a check that the layout
fits the pages.

Verified on hardware: short and long presses on both bank switches,
wrap-around at both ends, and GoTo and relative Bank commands, with the
display and button labels following the bank.

+++ firmware/Core/Inc/state_store.h
@@ -13,7 +13,7 @@
 #include <stdbool.h>
 
 // Load the last saved state. Returns false if nothing valid is stored.
-bool state_store_load(uint8_t *bank, uint8_t toggles[8], uint8_t long_toggles[8]);
+bool state_store_load(uint8_t *bank, uint32_t toggles[8], uint32_t long_toggles[8]);
 
 // Call whenever the bank or a toggle state changes. The save itself is
 // deferred until the state has been stable for a couple of seconds.
